+++ junit/clustering/KMeansTest.java
@@ -123,7 +123,7 @@ public class KMeansTest {
     }
 
     private void add(Dataset data, float x, float y) {
-        float[] values = { x, y };
+        double[] values = { x, y };
         SimpleInstance in = new SimpleInstance(values);
         data.addInstance(in);
     }
