+++ net/sf/javaml/core/InstanceTools.java
@@ -28,11 +28,11 @@ public class InstanceTools {
 
     
     public static Instance normalizeMidrange(double normalMiddle,double normalRange,Instance min,Instance max,Instance instance){
-        float[] out=new float[instance.size()];
+        double[] out=new double[instance.size()];
         for(int i=0;i<out.length;i++){
             double range=Math.abs(max.getValue(i)-min.getValue(i));
             double middle=Math.abs(max.getValue(i)+min.getValue(i))/2;
-            out[i]=(float)(((instance.getValue(i)-middle)/range)*normalRange+normalMiddle);
+            out[i]=((instance.getValue(i)-middle)/range)*normalRange+normalMiddle;
         }
         return new SimpleInstance(out,instance.getWeight(),instance.isClassSet(),instance.getClassValue());
     }
